kata-logger: use a larger ipc buffer
Use the magic seL4RPCOverMultiSharedData connector to allocate a page-size
shared memory buffer to each user of the LoggerInterface. This requires
a data copy between components but eliminates the 120 byte restriction
of the single ipc buffer used by seL4RPCCall. For now we limit log msgs
to 2048 bytes to reduce the possibility of walking off the end of the
thread stack (default allocation is 16KB) when handling a panic.

Change-Id: Ie1016f50d650cc066db8b4c2edc74a6c0f6c5e0c