Implements some std::io functions in kata_io
The ZMODEM implementation in lexxvir/zmodem calls functions added in
this change, which prepares for porting that library to no_std and
kata_io.

The core_io crate seemed to be a promising alternative, but an attempt
to use it found its build panicking with an "unknown compiler" error.
"It is intended that this crate be updated for every nightly":

https://github.com/jethrogb/rust-core_io/issues/15#issuecomment-477809487

Piecemeal adds of just the methods we really need is probably more
maintainable unless the pace of needing new methods increases.

Change-Id: Ib3c179ff974a6c537550e50f7401e1acbab06067