UART driver additional check for rx_empty
Throw-in: changes rx_semaphore post condition to be identical to the
condition being waited on.

Stragely, it can happen that STATUS.RXEMPTY is set even though
FIFO_STATUS.RXLVL == 0. For instance, pasting a ~100-character echo
command into the shell would reliably cause this condition, even if the
bound on the queue size was removed in OpenTitan_UART.cs.

It shouldn't be strictly necessary to also check RXEMPTY, but it is not
wrong to do even in a model where it would stay always in sync with
RXLVL.

This change lets us pass the "paste 100-char echo command" test, but it
doesn't get all the way to non-rate-limited and arbitrarily large ZMODEM
uploads. An uncommitted change that similarly adds and RXEMPTY check but
also goes back to polling instead of interrupts does work for "firehose"
ZMODEM uploads, so there is probably only a bit of subtlety remaining,
probably to do with rx_watermark handling.

Change-Id: I0a53c7092d4ad7eb5ece0d3e721241997870424b