sel4-sys: expose struct internals for generated shared types
Make the structs generated for shared types expose access to the internal
seL4_Words. This is important for capdl parsing that had no way to
convert CDL_CapData_Guard to an seL4_Word (for passing to the kernel).

While here do some minor cleanups of generated code (there are many
opportunities).

Change-Id: I179df84f3e863da4a82675a989c4faeb38f3f669