sel4-sys: fix x86_64 target
There is small caveat, as the scratch register had to be
changed (rbx -> r14) as llvm is using it _for something_.
In practice, it should not change anything, though.

sel4test test suites with MCS disabled are passing.
There is bug in mainline with doesn't allow to compile
seL4 kernel with -DLibSel4FunctionAttributes=public so
MCS enabled kernel cannot be tested.

Change-Id: I66271277ceb6d6242ddec1225590f8911f09e946