sel4-sys: fix aarch32 target
This fixes compiling and using sel4-sys for arm32 architecture --
tested on sabrelite seL4 taget.  There is some #[cfg(...)] ifology,
but that maybe could be partially moved to Cargo.toml using
feature dependency.

To fix TLS there is need to pass some rustc flags, but it seems that
it's impossible to do with build.rs, so currently it's only passed
by CMake when building for sel4test.

Additionally, the order of XML files had to be changed as the
syscalls identifiers were generated in wrong order.

Change-Id: I94d65284989b0e99d7b1843165ec88aa1a994be3