Fix seL4 slowdown by increasing ELF space.
In order to allocate a frame for the ELF, the seL4 object creation
routine scans through addresses, allocating frame by frame. This is very
slow if the ELF MMIO is sized 0x300000 as it requires a smaller seL4 object
(12 bits). Thus it allocates a 0x1000 frame over and over through the
space from 0x3c40_0000 to 0x4600_0000. Paradoxically if we instead
allocate 0x1000_0000 for the ELF, this scans the address range faster as
it needs a seL4 object of 22 bit size, so it scans the address space by
0x040_0000 increments instead.

Change-Id: Ic989e15eb0b2d65e26b141c7975d28cd611b5f79