handle additional condition syntax in generators
There is a change in mainline seL4, which changes the syntax
of conditional compilation condition:
https://github.com/seL4/seL4/commit/649b63d32f4d1d462aa756f0b5f92cb9ee6d89cc
file: libsel4/include/interfaces/sel4.xml

This is to adapt the existing generators to the newer seL4.
It adds a new case instead of replacing the existing one,
in order to keep it compatible with the older revisions of seL4.

Change-Id: I71148b10cb460b92e0555ee5a07ef15c2ae3eecc