kata-slot-allocator: CSpace slot allocator
Add an allocator for CSpace slot assignments. (This is generic enough
to work for any contiguous range of integer values.) The implementation
uses the bitvec crate to track slots and supports alloc+free of multiple
consecutive slots (to match up with ObjDescBundle). Allocators can be
setup global static with the usual 2-step empty+init api or use new.

Change-Id: I875c55e33bfb8915d01b273808d02875ba000020