Kata Timer Service
The Kata Timer Service multiplexes access to a single hardware timer,
providing up to 32 timers per client. Timers may either be single-shot
or periodic. Clients will connect with a seL4RPCCallSignal. When a timer
is triggered, the service notifies the client through the notification
attached to the RPC. The client then needs to call
`timer_completed_timers` to retrieve which timers are active and act
accordingly.

Change-Id: I2380524877823387dad7ecf48a72d67356853006