kata-os-model: reclaim rootserver resources
Reclaim memory assiged by the kernel to the rootserver to use in setting
up a running system. This is done by stashing the capabilities created
when processing the capDL specification in a "holding CNode" passed to
the MemoryManager (or whomever marks their CNode), and then revoking
UntypedMemory objects marked "tainted". The latter causes the kernel to
reclaim memory derived from the untypeds. The MemoryManager is otherwise
unaware (and does not use) the holding caps.

Specific changes:
- skip tainted untyped objects so capabilities it creates are not
  mixed with ones to be reclaimed
- revoke UntypedMemory caps marked tainted (in the MemoryManager)
- craft a CNode that holds rootserver-created caps and assign it
  to the MemoryManager (a la untyped objects)
- record "hidden capabilities" that need to be held in the dup cap
  table so the rootserver can sweep them into the holding CNode
- add CONFIG_NOISY_HANDOFF feature for spammy debug msgs

NB: the rootserver tries to move empty slots when sweeping caps into the
hold CNode; this spams the console with kernel msgs when CONFIG_PRINTING
is configured for the kernel

Change-Id: I34b87b09ac8708a8a2197670ce81daf20f5a6fd4