kata-security-coordinator: move test_mailbox to the real impl
Make the mailbox_api dependency optional so builds without a security
core work as intended. This means users of the fake lose the test_mailbox
shell command but given it was only a test vehicle it should be ok to
require configuring the sel4 feature.

As part of this update the "real impl" skeleton to the current traits.

Change-Id: I2a8628d316cca576d9c5dc579f099e16003a8f19