ProcessManager::sel4bundle: cleanup VSpace setup work
Replace the hacked up aarch64 support by something more modular & cleaner.
Split the construction of the ObjDesc set into objects common to all
platforms and arch-specific objects needed to setup the VSpace. As part
of this explicitly separate the indices into the dynamic_objs vector from
the CNode slot numbers where the objects land. This allows reordering &
each arch to have different sets of objects (e.g. 4-level page tables
on aarch64 vs 2-level on riscv32).

All the setup work is done using stack-based data which may be too much
of a burden on 64-bit archs but is generally preferred to using the heap
(SmallVec is your friend).

Calculate the vaddr for the (single) page table page we setup and note
the constraint this places on application size.

Update SmallVec use to the latest version.

Change-Id: Id9720e498442f64443ea32eddbd8ab14ee406dca