MlCoordinator: band-aid canceling a job with nothing loaded
If there are no images loaded after a job is canceled the image manager
tries to set the temp data region to zero length. This trips an assert
on debug builds and UB otherwise. Band-aid this by not setting the
region to a zero length and add an assert to catch this condition in a
release build.

Change-Id: Id27c968e6fbce421d95793c9806eb7a333c23d9b