SecurityCoordinator: reduce fakeimpl memory request size
The Fake support stores package contents in memory allocated by the
MemoryManager. For each Load* request it does a deep copy of this data
to MemoryManager allocated frames that are returned to the caller. These
objects were allocated with a single request with the assumption the
MemoryManager would break up the request as needed. But this is not
happening and won't happen immediately so instead use the page-at-a-time
logic used by the DebugConsole when retrieving package data from the
builtin cpio archive.

Bug: 243069315
Change-Id: Ie829ee68efa228ca25fdf2b17b98979e628d6ad8