cli: Give us better command line editing tools
This allows us to cursor around on the command line, insert new text,
and even edit previously entered text, a-la readline. It's not a
complete implementation of all of the emacs key bindings, so it's closer
to bash-style input, but it's far better than not being able to edit
properly.

Additionally filters out non-printable characters it doesn't recognize
(outside of the usual 7-bit ascii range) so we don't end up with
accidental garbage on the CLI.

Bug: 199515148
Change-Id: I92cc69b6cd3f9ffb5546acbd8c25dbc3485ec8a5