libcantrip: fix __sel4_ipc_buffer setup
The global ipc buffer is referenced using the __sel4_ipc_buffer TLS
symbol.  This symbol was setup in both assembler and Rust (the latter
initializing it to zero). Somehow this worked on riscv32 but on aarch64
the Rust init happened last causing any access to the ipc buffer to
fault. This was further masked by the ipc buffer not being used for
seL4_DebugPutChar syscalls (args passed only in registers). Remove
the Rust setup so there is no confusion about who is responsible for
setup.

While here make the riscv32 & aarch64 crt0 code more consistent and
trim the static TLS segment (symbol "_tls") to just what's needed;
this shaves a page off the bss segment.

Change-Id: Ieb9de0132d126e890765317aefacb586fdf03092