libcantrip: replace cantrip-os-logger with an sdk-specific version
cantrip-os-logger's use of a weak symbol for logger_log worked on
riscv23 but not on aarch64. Given weak symbol linkage is an unstable
feature switch to a private version that connects directly to sdk_log.
This also simplifies the logic and enables changing (in the future)
to marshaling parameters directly into the SDK rpc shared page.

Change-Id: I191457f124c7fe1f44f0e88d260f418e04bfcee4