Camkes: band-aid capability leakage in reply msgs
We blindly enable attaching a capability to reply messages for some
connectors (until we replace the C templates). This means we must be
careful about clearing the capability field in the ipc buffer so that
values do not leak into replies causing unintended spammage.  This was
done for the client-side of rpc's but not entirely for the server-side
causing unintended behaviour. Add Camkes::cleanup_request_cap() to return
an RAII cleanup helper that clears the request/reply capability in the
ipc buffer and use it in all components that need it.

While here do some minor tidying:
- hoist an assert in cantrip-memory-interface support
- correct a comment typo
- remove Camkes::set_reply_cap as it was unsafe to use

Change-Id: Id0a562f5e3cf59719ed693f390ec226548e91638