cantrip: move builtins from DebugConsole to SecurityCoordinator
The builtins archive was wrongly located in the DebugConsole. It was
intended to be in eFLASH in the SecurityCore (SC) and accessed via the
SecurityCoordinator but for now move it as close as possible--into the
SecurityCoordinator fake (in the future a real implementation can access
the data via the SC's mailbox api).

With builtins in the SecurityCoordinator it is possible automatically
instantiate applications in response to a LoadApplication request and
the ProcessManager now auto-installs a bundle when asked to Start
an application. This makes it possible to operate without install &
uninstall shell commands (see below)--there is a small space overhead
in ProcessManager for not uninstall'ing a bundle.

The DebugConsole now uses LoadApplication to read .repl files with
commands and implements autostart.repl using the source command.

With the above changes a release build now runs the shodan_boot robot
script in 4MiB of memory (there is ~2.5MiB of memory available for
applications & models).

Specific changes:
- move the cpio_archive dataport to the SecurityCoordinator
- change autostart behaviour too use a source command to consolidate
  .repl file handling in one place
- add a dynamic_load_support feature to cantrip-shell (default off)
- zmodem support is now required only with dynamic_load_support
- cpio support is no longer part of the base system (only needed by
  the fake SecurityCoordinator)
- add a packages shell command that displays builtins + dynamically
  installed packages; this is implemented with a new
  cantrip_security_get_packages api
- add a -q option to the source command to not complain if the
  requested file is not found
- feature-gate the shell install & uninstall commands under
  "dynamic_load_support" which is disabled by default.
- bump the capacity of the fake SecurityCoordinator HashMap to 10;
  it should be large enough to hold all builtins (though it should
  spill to the heap as needed)
- accept both "foo" and "foo.app" (for example) the fake
  SecurityCoordinator for backwards compatibility; this is a bit
  expensive but likely ok since it's isolated to the fake

NB: Defaulting dynamic_load_support to off means "sealed system images"
  are now the default. Disabling dynamic loading also enables stripping
  dependent functinality in other services (e.g. ProcessManager).

Fix: 267187709

Change-Id: Ie750df5a99ce2640606d62138401e6d53ce0d53f