MailboxDriver: move irq processing to the control thread
Switch from processing IRQ's in dedicated threads to processing most
IRQ's in the control thread (that is otherwise idle). The rtirq is
left with a dedicated thread for now as it defers ack to the api thread
and that doesn't fit into the shared irq model (atm) This saves ~20KB
by eliminating 2x threads: TCB+SC (1K), stack (4K), ipc_buffer (4K).

Change-Id: I0478475103ff59b5649f18ab89f07dbf161a7bbb