processmanager: Add in a way to get a process exit status
This is the first step to giving us a DebugConsole "waitfor" command.
With it, we get a beginning command that just polls ProcessManager
for the status of a bundle ID. Due to the way bundles are "installed"
from the builtins, the status command actually fails to find the
bundles before they're started for the first time.

Change-Id: I4deab3bca839de5e069f32fd1f33bbe7b4822bb7