shodan: unify camkes platform configuration
Leverage camkes' hidden use of cpp to include a platform config that
has hardware configuration like the physical addresses of devices and
irq assignments. The hardware configuration uses #defines generated
from the RTL to ensure a single SOT.

This also requires a platform.camkes for rpi3/bcm2837 that specifies the
address & size of the cpio bundle.

Bug: 276823440

Change-Id: I73df3fc7177aba060c81f5aabd9ad4ba0b739ad3