camkes/irq: generalize irq handling by removing static_irq_thread
An irq thread is just an interface thread with a boilerplate run method.
Instead of baking this into camkes-generated code just create interface
threads and extend the irq support to simplify generating the boilerplate
code. This gives more flexibility in writing irq handlers and fixes
the awkwardness required to supply init() methods.

Change-Id: I9a2d18764d8c48d09f120a0687c161e3043160e0