MlCoordinator: cleanup some aspects of periodic jobs
When a periodic job runs more slowly than the associated timer (which
is not tied to the duty cycle of the job) distinguish between a timer
expiring while the job is executing and while it is sitting in the
execution queue waiting to be scheduled. Both can be true if a job takes
longer than 2x the periodic timer interval so count the events separately
and generate only one message. Make the messages more informative
(hopefully).

While here, tweak the finish interrupt handler to reset the vector core
before starting any next model. On springbok a reset is a nop but on
kelvin it is unclear exactly what it does; but starting the core before
immediately reset'ing it seems wrong.

Change-Id: Ia29ee1cbe375e55d9f750460cf7f5581b7811f4b