SDKRuntime/i2s: double-buffer recording
Add a second 2K sample buffer with one buffer used to collect samples
from the FIFO with the other drained through audio_record_collect calls.
Buffer flips happen in the rx_watermark handler when the record buffer
is full and the collect buffer is empty (so there are no partial buffers
to deal with).

Change-Id: Ia4b7c8d11b0db218377a324c1ad7237010a15e7e